+++ src/PufferfishSparseIndex.cpp
@@ -351,7 +351,7 @@ std::string PufferfishSparseIndex::getSeqStr(size_t globalPos, size_t length, bo
  	word = seq_.get_int(2*globalPos, 2*validLength);
 	globalPos += validLength;
 	if (isFw)
-	  for(int i = 0; i < 2*validLength ;i+=2){
+	  for(uint64_t i = 0; i < 2*validLength ;i+=2){
 		base = (word >> i) & 0x03;
 	    switch(base){
     	case 0:
@@ -369,7 +369,7 @@ std::string PufferfishSparseIndex::getSeqStr(size_t globalPos, size_t length, bo
     	}
     }
 	else
-    for(int i = 0; i < 2*validLength ;i+=2){
+    for(uint64_t i = 0; i < 2*validLength ;i+=2){
       base = (word >> i) & 0x03;
       switch(base){
       case 0:
